main.lua in mehrere Skriptdateien aufgeteilt.
