+++ worlds/plugins/unitopia/unitopia.xml
@@ -203,7 +203,7 @@
 
 <script>
 <![CDATA[
-require("unitopia.main")
+require("unitopia.main.main")
 ]]>
 </script>
 </muclient>
